assets/requirements-backend.txt
# Helix Collective v15.3 Dual Resonance - Backend Requirements (Railway)

# Core Framework & Web Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
jinja2==3.1.2  # Template engine for web dashboard
aiofiles==23.2.1  # For async file operations (moved up)

# Discord Integration
discord.py==2.4.0

# AI Model APIs
anthropic==0.39.0
google-generativeai==0.8.3
openai==1.54.3

# Data & State Management
redis[hiredis]==5.2.0
asyncpg==0.30.0
sqlalchemy==2.0.23

# MEGA Sync - pycryptodome + tenacity (mega.py installed separately with --no-deps in Dockerfile)
pycryptodome
tenacity  # Required by mega.py for retry logic

# Data Science & Analytics
numpy==1.26.4
pandas==2.1.3
matplotlib==3.8.0
scikit-learn
prophet

# Audio & Ritual Engine
pydub==0.25.1

# API Clients
httpx==0.28.0
aiohttp==3.11.0
requests==2.31.0

# Monitoring & Observability
sentry-sdk[fastapi]==2.19.0

# Utilities
pyyaml==6.0.1
python-multipart==0.0.6
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3

# Notion Integration
notion-client==2.5.0

Pillow==10.1.0  # For Grok's fractal icon generation

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.28.0  # For testing async HTTP (already listed above)

# Database Migrations
alembic==1.13.1

# Optional Storage Dependencies (install separately if needed)
# webdav3  # Not available for Python 3.11+ - install manually if using Nextcloud
# boto3    # Optional - install if using Backblaze B2 storage
